Share core logic
